requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
boto3==1.38.32
botocore==1.38.32
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
fasta2a==0.2.15
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
firecrawl-py==2.8.0
frozenlist==1.6.2
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.19.0
griffe==1.7.3
groq==0.26.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.64
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
logfire-api==3.18.0
markdown-it-py==3.0.0
mcp==1.9.3
mdurl==0.1.2
mistralai==1.8.1
multidict==6.4.4
nest-asyncio==1.6.0
openai==1.84.0
opentelemetry-api==1.34.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
prompt_toolkit==3.0.51
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-ai==0.2.15
pydantic-ai-slim==0.2.15
pydantic-evals==0.2.15
pydantic-graph==0.2.15
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
